chore(benches): create super naiv benchmarks
This is a benchmark which is supposed to show that using slotmap in a
new graph lib implementation should give us some performance benefits
over indexmap which is currently used in petgraph.

The benchmarks should be extended to other areas like iteration, access
etc. and they really need to be improved by someone with more
benchmarking knowledge (or me after diving a bit more into benchmarking)

Authored-by: Aviac <aviac@mailbox.org>